@@ -1,7 +1,10 @@
-abstract class MnemonicEvent {}
+import 'package:equatable/equatable.dart';
+
+abstract class MnemonicEvent extends Equatable {
+  @override
+  List<Object> get props => [];
+}
 
 class MnemonicGenerationPressed extends MnemonicEvent {}
-
-class MnemonicRecoveryInserted extends MnemonicEvent {}
 
 class MnemonicConfirmed extends MnemonicEvent {}
