@@ -1,32 +1,28 @@
 import 'package:equatable/equatable.dart';
 
 abstract class MnemonicState extends Equatable {
-  const MnemonicState({this.mnemonic});
-  final String? mnemonic;
+  const MnemonicState({this.mnemonic = ""});
+  final String mnemonic;
+  //final bool isWrittenDownByUser;
 
   @override
   List<Object?> get props => [mnemonic];
 }
 
-class MnemonicInitial extends MnemonicState {
-  const MnemonicInitial();
+class MnemonicInitial extends MnemonicState {}
+
+class MnemonicSuccess extends MnemonicState {
+  const MnemonicSuccess({required String mnemonic}) : super(mnemonic: mnemonic);
 }
 
-class MnemonicGenerationInProgress extends MnemonicState {
-  const MnemonicGenerationInProgress();
-}
+class MnemonicFailure extends MnemonicState {}
 
-class MnemonicGenerationSuccess extends MnemonicState {
-  const MnemonicGenerationSuccess({required String mnemonic})
+class MnemonicStoreSuccess extends MnemonicState {
+  const MnemonicStoreSuccess({required String mnemonic})
       : super(mnemonic: mnemonic);
 }
 
-class MnemonicGenerationFailure extends MnemonicState {}
-
-class MnemonicInputSuccess extends MnemonicState {}
-
-class MnemonicInputFailure extends MnemonicState {}
-
-class MnemonicStoreSuccess extends MnemonicState {}
-
-class MnemonicStoreFailure extends MnemonicState {}
+class MnemonicStoreFailure extends MnemonicState {
+  const MnemonicStoreFailure({required String mnemonic})
+      : super(mnemonic: mnemonic);
+}
